Refactor PEM/OE/ADAM for KISS, DRY, and dead code elimination
Architecture improvements based on code review:

1. DRY - Covariance estimation (PEM & OE):
   - Extracted _estimate_covariance() to SysIdAlgBase
   - Accepts optional reshape_func for PolynomialModel compatibility
   - Removes 60+ lines of duplicate code

2. DRY - Gradient computation in ADAM:
   - Consolidated _compute_gradient_finite_small_epsilon and _compute_gradient_finite_numba
   - Single _compute_gradient_finite method using vectorized NumPy
   - Removed Numba remnants (dead code)

3. KISS - Benchmark code:
   - Moved benchmark_derivative_methods and print_benchmark_results
   - New location: scripts/benchmark_gradients.py
   - Reduces pem.py by ~200 lines of analytical code

4. KISS - Nested functions in OE._ident:
   - Extracted objective_analytical, objective_finite, objective_complex
   - Now class methods: _objective_analytical, _objective_finite, _objective_complex
   - Parameters stored as instance variables (_oe_nb, _oe_nf_full, _oe_nk, _oe_mod_Ts)
   - Massively improves readability of _ident method

5. Dead Code:
   - Removed all Numba-related code (imports, decorators, fallback)
   - Removed _compute_gradient_finite_numba and _compute_gradient_finite_small_epsilon

Net result: pem.py reduced by ~455 lines, code is more maintainable and follows KISS/DRY principles.
