+++ src/llsi/pem.py
@@ -101,45 +101,8 @@ class PEM(SysIdAlgBase):
         # Update model with optimized parameters
         mod.reshape(res.x)
 
-        # --- Corrected Covariance Estimation ---
-        # 1. Calculate the Jacobian of the residuals J_res (N x P)
-        # We need manual finite differences because approx_fprime is for scalar outputs
-
-        theta_opt = res.x
-        n_params = len(theta_opt)
-        n_samples = self.y.size
-        epsilon = 1e-8
-
-        J_res = np.zeros((n_samples, n_params))
-        y_nominal = mod.simulate(self.u).ravel()
-
-        for i in range(n_params):
-            theta_perturbed = theta_opt.copy()
-            theta_perturbed[i] += epsilon
-
-            # Temporarily set model params to perturbed values
-            mod.reshape(theta_perturbed)
-            y_perturbed = mod.simulate(self.u).ravel()
-
-            # Jacobian column i = - dy_hat / dtheta
-            # Residual = y - y_hat => d(Res)/dtheta = - dy_hat/dtheta
-            J_res[:, i] = (y_nominal - y_perturbed) / epsilon
-
-        # Restore model to optimum
-        mod.reshape(theta_opt)
-
-        # 2. Estimate Variance of residuals
-        residuals = self.y.ravel() - y_nominal
-        sigma2 = np.sum(residuals**2) / (n_samples - n_params)
-
-        # 3. Compute Covariance: Cov = sigma^2 * (J^T J)^-1
-        H_approx = J_res.T @ J_res
-
-        try:
-            mod.cov = sigma2 * np.linalg.inv(H_approx)
-        except np.linalg.LinAlgError:
-            # Fallback for ill-conditioned matrices
-            mod.cov = sigma2 * np.linalg.pinv(H_approx)
+        # Estimate covariance matrix using finite differences
+        self._estimate_covariance(mod, res.x, epsilon=1e-8)
 
         return mod
 
@@ -226,8 +189,10 @@ class ADAM(SysIdAlgBase):
             - 'finite': Uses scipy.optimize.approx_fprime (slower, less accurate).
         """
         if self.derivative_method == "complex":
-            return self._compute_gradient_finite_small_epsilon(x, y_batch, u_batch)
+            # Use finite differences with small epsilon (1e-8) for better accuracy
+            return self._compute_gradient_finite(x, y_batch, u_batch)
         else:
+            # Use scipy's approx_fprime (slower, less accurate)
             return self._compute_gradient_finite(x, y_batch, u_batch)
 
     def _compute_gradient_finite(self, x: np.ndarray, y_batch: np.ndarray, u_batch: np.ndarray) -> np.ndarray:
@@ -243,18 +208,29 @@ class ADAM(SysIdAlgBase):
 
 
 
-    def _compute_gradient_finite_numba(self, x: np.ndarray, y_batch: np.ndarray, u_batch: np.ndarray) -> np.ndarray:
+    def _compute_gradient_finite(self, x: np.ndarray, y_batch: np.ndarray, u_batch: np.ndarray) -> np.ndarray:
         """
-        Finite difference gradient computation with pre-allocated arrays.
+        Compute gradient using forward finite differences with epsilon=1e-8.
 
-        Uses forward differences with vectorized NumPy operations for the final
-        gradient computation. Numba was removed as it provided no benefit for
-        this trivial operation (vectorized NumPy is already C-speed).
+        This is the consolidated gradient computation method for ADAM optimizer.
+        It uses a small epsilon for better accuracy than standard finite differences
+        (1e-4 to 1e-6) while remaining numerically stable for real-valued simulations.
+
+        Note: For true complex step, use the OE class with derivative_method='complex',
+        which supports complex coefficients via oe_simulate.
+
+        Args:
+            x: Parameter vector.
+            y_batch: Output batch.
+            u_batch: Input batch.
+
+        Returns:
+            Gradient vector.
         """
         epsilon = 1e-8
         n_params = len(x)
 
-        # Pre-compute nominal loss
+        # Pre-compute nominal loss (includes regularization via compute_loss)
         nominal_loss = self.compute_loss(x, y_batch, u_batch)
 
         # Pre-allocate array for perturbed losses
@@ -266,63 +242,8 @@ class ADAM(SysIdAlgBase):
             x_perturbed[i] += epsilon
             losses[i] = self.compute_loss(x_perturbed, y_batch, u_batch)
 
-        # Vectorized gradient computation (faster than Numba for this simple operation)
+        # Vectorized gradient computation (NumPy is already C-speed for this)
         grad = (losses - nominal_loss) / epsilon
-
-        return grad
-
-    def _compute_gradient_finite_small_epsilon(self, x: np.ndarray, y_batch: np.ndarray, u_batch: np.ndarray) -> np.ndarray:
-        """
-        Compute gradient using finite differences with a very small epsilon.
-
-        Note: This was previously mislabeled as "complex step" but actually uses
-        a real perturbation. The small epsilon (1e-8) provides better accuracy
-        than standard finite differences (1e-4 to 1e-6) but is numerically stable
-        for real-valued simulations. For true complex step, use OE.objective_complex
-        which supports complex coefficients via oe_simulate.
-
-        Uses Numba JIT compilation for acceleration when available (n_params >= 20).
-        """
-        epsilon = 1e-8  # Small perturbation for better accuracy (not complex step!)
-        n_params = len(x)
-
-        # Use Numba-accelerated version for larger problems
-        if n_params >= 20:
-            return self._compute_gradient_finite_numba(x, y_batch, u_batch)
-
-        grad = np.zeros_like(x)
-
-        if self.model is None:
-            raise RuntimeError("Model not initialized.")
-
-        # Pre-compute nominal loss and output
-        self.model.reshape(x)
-        y_nominal = self.model.simulate(u_batch)
-        nominal_loss = LTIModel.SSE(y_batch - y_nominal)
-
-        # Add nominal regularization
-        if self.lambda_l1 > 0:
-            nominal_loss += self.lambda_l1 * np.sum(np.abs(x))
-        if self.lambda_l2 > 0:
-            nominal_loss += self.lambda_l2 * (x.T @ x)
-
-        for i in range(n_params):
-            # Perturb parameter i with small real epsilon
-            x_perturbed = x.copy()
-            x_perturbed[i] += epsilon
-
-            self.model.reshape(x_perturbed)
-            y_perturbed = self.model.simulate(u_batch)
-            perturbed_loss = LTIModel.SSE(y_batch - y_perturbed)
-
-            # Add regularization terms for perturbed parameters
-            if self.lambda_l1 > 0:
-                perturbed_loss += self.lambda_l1 * np.sum(np.abs(x_perturbed))
-            if self.lambda_l2 > 0:
-                perturbed_loss += self.lambda_l2 * (x_perturbed.T @ x_perturbed)
-
-            # Forward finite difference
-            grad[i] = (perturbed_loss - nominal_loss) / epsilon
 
         return grad
 
@@ -435,6 +356,80 @@ class OE(PEM):
 
         self._math = _math
 
+    def _objective_analytical(self, theta: np.ndarray) -> tuple[float, np.ndarray]:
+        """
+        Objective function for OE identification with analytical gradients and overflow protection.
+        Returns (cost, gradient) for scipy.optimize.minimize.
+        """
+        cost, grad = self._math.oe_cost_and_gradient(
+            theta, self.u.ravel(), self.y.ravel(), self._oe_nb, self._oe_nf_full, self._oe_nk
+        )
+
+        # Protect against unstable simulations (poles outside unit circle)
+        # Replace NaN/inf with massive penalty to force optimizer to retreat
+        if not np.isfinite(cost):
+            cost = 1e300
+            grad = np.zeros_like(grad)
+        elif not np.all(np.isfinite(grad)):
+            # Sanitize gradients to prevent optimizer math errors
+            grad = np.nan_to_num(grad, nan=0.0, posinf=1e10, neginf=-1e10)
+
+        return cost, grad
+
+    def _objective_finite(self, theta: np.ndarray) -> float:
+        """
+        Objective function using finite differences for gradient computation.
+        """
+        from .polynomialmodel import PolynomialModel
+
+        mod_temp = PolynomialModel(
+            a=np.concatenate(([1.0], theta[self._oe_nb :])),
+            b=theta[: self._oe_nb],
+            nk=self._oe_nk,
+            Ts=self._oe_mod_Ts,
+        )
+        y_hat = mod_temp.simulate(self.u)
+        sse = LTIModel.SSE(self.y - y_hat)
+        # Handle numerical instability
+        return float(np.nan_to_num(sse, nan=1e300))
+
+    def _objective_finite_wrapper(self, theta: np.ndarray) -> tuple[float, np.ndarray]:
+        """Wrapper for _objective_finite to return (cost, grad) tuple with zero gradient."""
+        return (self._objective_finite(theta), np.zeros_like(theta))
+
+    def _objective_complex(self, theta: np.ndarray) -> tuple[float, np.ndarray]:
+        """
+        Objective function using complex step method for gradient computation.
+        Uses direct simulation with complex coefficients via oe_simulate.
+        """
+        cost = self._objective_finite(theta)
+        epsilon = 1e-20
+        grad = np.zeros_like(theta)
+
+        for i in range(len(theta)):
+            # Perturb parameter i with complex step
+            theta_perturbed = theta.copy().astype(np.complex128)
+            theta_perturbed[i] = theta[i] + epsilon * 1j
+
+            # Simulate with complex coefficients using oe_simulate
+            b_complex = theta_perturbed[: self._oe_nb]
+            f_complex = np.concatenate((np.array([1.0 + 0j]), theta_perturbed[self._oe_nb :]))
+            y_hat_complex = self._math.oe_simulate(self.u.ravel(), b_complex, f_complex, self._oe_nk)
+
+            # Compute cost with complex output (do NOT cast to float!)
+            cost_complex = np.sum((self.y.ravel() - y_hat_complex) ** 2)
+
+            grad[i] = np.imag(cost_complex) / epsilon
+
+        # Protect against unstable simulations
+        if not np.isfinite(cost):
+            cost = 1e300
+            grad = np.zeros_like(grad)
+        elif not np.all(np.isfinite(grad)):
+            grad = np.nan_to_num(grad, nan=0.0, posinf=1e10, neginf=-1e10)
+
+        return cost, grad
+
     def _ident(self, order: Union[int, tuple[int, ...]]) -> LTIModel:
         """
         Identify OE model using analytical gradients.
@@ -492,95 +487,25 @@ class OE(PEM):
         nf_full = nf_oe + 1
         n_params = nb_oe + nf_oe  # Total number of parameters
 
-        # Define objective function with analytical gradient and overflow protection
-
-        def objective_analytical(theta: np.ndarray) -> tuple[float, np.ndarray]:
-            """
-            Objective function for OE identification with analytical gradients and overflow protection.
-            Returns (cost, gradient) for scipy.optimize.minimize.
-            """
-            cost, grad = self._math.oe_cost_and_gradient(theta, self.u.ravel(), self.y.ravel(), nb_oe, nf_full, nk)
-
-            # Protect against unstable simulations (poles outside unit circle)
-            # Replace NaN/inf with massive penalty to force optimizer to retreat
-            if not np.isfinite(cost):
-                cost = 1e300
-                grad = np.zeros_like(grad)
-            elif not np.all(np.isfinite(grad)):
-                # Sanitize gradients to prevent optimizer math errors
-                grad = np.nan_to_num(grad, nan=0.0, posinf=1e10, neginf=-1e10)
-
-            return cost, grad
-
-        # Define objective function using finite differences (for fallback or explicit selection)
-
-        def objective_finite(theta: np.ndarray) -> float:
-            """
-            Objective function using finite differences for gradient computation.
-            """
-            mod_temp = PolynomialModel(
-                a=np.concatenate(([1.0], theta[nb_oe:])),
-                b=theta[:nb_oe],
-                nk=nk,
-                Ts=mod.Ts,
-            )
-            y_hat = mod_temp.simulate(self.u)
-            sse = LTIModel.SSE(self.y - y_hat)
-            # Handle numerical instability
-            return float(np.nan_to_num(sse, nan=1e300))
+        # Store parameters as instance variables for use in objective functions
+        self._oe_nb = nb_oe
+        self._oe_nf_full = nf_full
+        self._oe_nk = nk
+        self._oe_mod_Ts = mod.Ts
 
         # Select derivative method
         if derivative_method == "analytical":
-            objective = objective_analytical
+            objective = self._objective_analytical
             use_analytical = True
         elif derivative_method == "finite":
-
-            def objective_finite_wrapper(theta: np.ndarray) -> tuple[float, np.ndarray]:
-                return (objective_finite(theta), np.zeros_like(theta))
-
-            objective = objective_finite_wrapper
+            objective = self._objective_finite_wrapper
             use_analytical = False
         elif derivative_method == "complex":
-            # Complex step method - use very small epsilon
-
-            def objective_complex(theta: np.ndarray) -> tuple[float, np.ndarray]:
-                """
-                Objective function using complex step method for gradient computation.
-                Uses direct simulation with complex coefficients via oe_simulate.
-                """
-                cost = objective_finite(theta)
-                epsilon = 1e-20
-                grad = np.zeros_like(theta)
-
-                for i in range(len(theta)):
-                    # Perturb parameter i with complex step
-                    theta_perturbed = theta.copy().astype(np.complex128)
-                    theta_perturbed[i] = theta[i] + epsilon * 1j
-
-                    # Simulate with complex coefficients using oe_simulate
-                    b_complex = theta_perturbed[:nb_oe]
-                    f_complex = np.concatenate((np.array([1.0 + 0j]), theta_perturbed[nb_oe:]))
-                    y_hat_complex = self._math.oe_simulate(self.u.ravel(), b_complex, f_complex, nk)
-
-                    # Compute cost with complex output (do NOT cast to float!)
-                    cost_complex = np.sum((self.y.ravel() - y_hat_complex) ** 2)
-
-                    grad[i] = np.imag(cost_complex) / epsilon
-
-                # Protect against unstable simulations
-                if not np.isfinite(cost):
-                    cost = 1e300
-                    grad = np.zeros_like(grad)
-                elif not np.all(np.isfinite(grad)):
-                    grad = np.nan_to_num(grad, nan=0.0, posinf=1e10, neginf=-1e10)
-
-                return cost, grad
-
-            objective = objective_complex
+            objective = self._objective_complex
             use_analytical = True
         else:
             self.logger.warning(f"Unknown derivative_method '{derivative_method}'. Using 'analytical'.")
-            objective = objective_analytical
+            objective = self._objective_analytical
             use_analytical = True
 
         # Get minimizer settings
@@ -639,351 +564,16 @@ class OE(PEM):
             Ts=mod.Ts,
         )
 
-        # Compute covariance matrix using the Jacobian of residuals
-        # We need to compute J_res = -dy_hat/dtheta (residual = y - y_hat, so d(residual)/dtheta = -dy_hat/dtheta)
-        n_params = len(theta_opt)
-        n_samples = len(self.y)
-        epsilon = 1e-8
+        # Define a reshape function for PolynomialModel that maps theta to (a, b)
+        def reshape_polynomial(theta: np.ndarray) -> None:
+            mod.a = np.concatenate(([1.0], theta[nb_oe:]))
+            mod.b = theta[:nb_oe]
 
-        J_res = np.zeros((n_samples, n_params))
-        y_nominal = mod.simulate(self.u).ravel()
-
-        for i in range(n_params):
-            theta_perturbed = theta_opt.copy()
-            theta_perturbed[i] += epsilon
-
-            # Create temporary model with perturbed parameters
-            mod_perturbed = PolynomialModel(
-                a=np.concatenate(([1.0], theta_perturbed[nb_oe:])),
-                b=theta_perturbed[:nb_oe],
-                nk=nk,
-                Ts=mod.Ts,
-            )
-            y_perturbed = mod_perturbed.simulate(self.u).ravel()
-
-            # Jacobian column: d(residual)/dtheta_i = d(y - y_hat)/dtheta_i = -dy_hat/dtheta_i
-            J_res[:, i] = (y_nominal - y_perturbed) / epsilon
-
-        # Estimate variance of residuals
-        residuals = self.y.ravel() - y_nominal
-        sigma2 = np.sum(residuals**2) / (n_samples - n_params)
-
-        # Compute covariance: Cov = sigma^2 * (J^T J)^-1
-        H_approx = J_res.T @ J_res
-        try:
-            mod.cov = sigma2 * np.linalg.inv(H_approx)
-        except np.linalg.LinAlgError:
-            mod.cov = sigma2 * np.linalg.pinv(H_approx)
+        # Estimate covariance matrix using finite differences
+        self._estimate_covariance(mod, theta_opt, epsilon=1e-8, reshape_func=reshape_polynomial)
 
         return mod
 
     @staticmethod
     def name() -> str:
         return "oe"
-
-
-def benchmark_derivative_methods(
-    data: Optional[SysIdData] = None,
-    y_name: Optional[Union[str, list[str]]] = None,
-    u_name: Optional[Union[str, list[str]]] = None,
-    order: Union[int, tuple[int, ...]] = 2,
-    n_runs: int = 5,
-    n_params_list: Optional[list[int]] = None,
-    loss_function: Optional[Callable] = None,
-    mod: Optional[LTIModel] = None,
-) -> dict[str, Any]:
-    """
-    Benchmark finite differences vs. complex step derivative methods.
-
-    This function compares the speed and accuracy of two numerical differentiation
-    methods for computing gradients in optimization problems.
-
-    Args:
-        data: System identification data (optional, for real simulation-based benchmarking).
-        y_name: Output channel name(s) (optional, required if data is provided).
-        u_name: Input channel name(s) (optional, required if data is provided).
-        order: Model order for initialization (optional, used if data is provided).
-        n_runs: Number of benchmark runs per configuration.
-        n_params_list: List of parameter counts to test.
-        loss_function: Custom loss function for benchmarking. If None and no data/mod,
-                      uses a simple quadratic loss. Signature: f(x) -> float.
-        mod: LTIModel instance (optional). If provided, uses mod.simulate for realistic
-             benchmarking. Requires y_name, u_name, and data to be set.
-
-    Returns:
-        Dictionary with benchmark results:
-        {
-            'methods': ['finite', 'complex'],
-            'n_params': [10, 50, 100],
-            'times': {'finite': [...], 'complex': [...]},
-            'errors': {'finite': [...], 'complex': [...]},
-        }
-
-    Notes:
-        - Complex step is generally FASTER due to fewer function evaluations
-          (1 vs 2 per parameter) and ALWAYS more accurate (error ~1e-16 vs ~1e-7).
-        - For very complex loss functions with expensive evaluations (e.g., simulations),
-          the reduced number of evaluations in complex step typically outweighs
-          the overhead of complex arithmetic. However, in practice, complex arithmetic
-          in NumPy is ~2x slower than real arithmetic, so the speedup may be less than
-          theoretical for simple functions.
-        - When using a real model (mod argument), the 'complex' method will only work
-          if the model's simulate method supports complex parameters (e.g., OE with
-          oe_simulate). Otherwise, it falls back to finite differences with small epsilon.
-
-    Example:
-        >>> import numpy as np
-        >>> from llsi.pem import benchmark_derivative_methods, print_benchmark_results
-        >>> # Simple benchmark with quadratic loss
-        >>> results = benchmark_derivative_methods(n_params_list=[10, 50, 100])
-        >>> print_benchmark_results(results)
-        >>>
-        >>> # With custom loss function
-        >>> def my_loss(x):
-        ...     return np.sum(np.sin(x) ** 2)
-        >>> results = benchmark_derivative_methods(loss_function=my_loss)
-        >>>
-        >>> # With real simulation (requires data, y_name, u_name)
-        >>> from llsi.sysiddata import SysIdData
-        >>> data = SysIdData(...)
-        >>> results = benchmark_derivative_methods(
-        ...     data=data, y_name='y', u_name='u', order=2, n_params_list=[10]
-        ... )
-    """
-    import time
-
-    # Initialize n_params_list with default if None
-    if n_params_list is None:
-        n_params_list = [10, 50, 100]
-
-    results = {
-        "methods": ["finite", "complex"],
-        "n_params": n_params_list,
-        "times": {"finite": [], "complex": []},
-        "errors": {"finite": [], "complex": []},
-    }
-
-    # Check if we should use simulation-based benchmarking
-    use_simulation = mod is not None and data is not None and y_name is not None and u_name is not None
-
-    if use_simulation:
-        # Extract y and u from data
-        y = data.get_output(y_name) if isinstance(y_name, str) else data.get_output(y_name)
-        u = data.get_input(u_name) if isinstance(u_name, str) else data.get_input(u_name)
-
-        # Vectorize model parameters for benchmarking
-        x0 = mod.vectorize()
-        n_params_model = len(x0)
-
-        # Override n_params_list with model's parameter count if not provided
-        if n_params_list is None or max(n_params_list) > n_params_model:
-            n_params_list = [n_params_model]
-
-        # True gradient: Not analytically available, so we use a high-accuracy finite difference as reference
-        # Compute reference gradient with very small epsilon
-        epsilon_ref = 1e-10
-        grad_ref = np.zeros(n_params_model)
-        loss_nominal = LTIModel.SSE(y - mod.simulate(u))
-
-        for i in range(n_params_model):
-            x_perturbed = x0.copy()
-            x_perturbed[i] += epsilon_ref
-            mod.reshape(x_perturbed)
-            loss_perturbed = LTIModel.SSE(y - mod.simulate(u))
-            grad_ref[i] = (loss_perturbed - loss_nominal) / epsilon_ref
-
-        mod.reshape(x0)  # Reset model
-
-        for n_params in n_params_list:
-            if n_params > n_params_model:
-                # Skip if n_params exceeds model parameters
-                results["times"]["finite"].append(np.nan)
-                results["errors"]["finite"].append(np.nan)
-                results["times"]["complex"].append(np.nan)
-                results["errors"]["complex"].append(np.nan)
-                continue
-
-            x_test = x0[:n_params].copy()
-            true_grad = grad_ref[:n_params]
-
-            # Benchmark finite differences
-            finite_times = []
-            finite_errors = []
-
-            for _ in range(n_runs):
-                start = time.perf_counter()
-
-                # Use scipy's approx_fprime with simulation-based loss
-                # Use a factory function to avoid closure issues with n_params
-                def make_sim_loss(n_params_bound: int) -> Callable[[np.ndarray], float]:
-                    def _sim_loss(x_test_inner: np.ndarray) -> float:
-                        mod.reshape(np.concatenate([x_test_inner, x0[n_params_bound:]]))
-                        return float(LTIModel.SSE(y - mod.simulate(u)))
-                    return _sim_loss
-
-                grad_finite = scipy.optimize.approx_fprime(x_test, make_sim_loss(n_params), epsilon=1e-8)
-
-                elapsed = time.perf_counter() - start
-                finite_times.append(elapsed)
-                finite_errors.append(float(np.linalg.norm(grad_finite - true_grad)))
-
-            results["times"]["finite"].append(np.median(finite_times))
-            results["errors"]["finite"].append(np.median(finite_errors))
-
-            # Benchmark complex step (only works if model supports complex parameters)
-            complex_times = []
-            complex_errors = []
-
-            try:
-                for _ in range(n_runs):
-                    start = time.perf_counter()
-
-                    epsilon = 1e-20
-                    grad_complex = np.zeros_like(x_test)
-                    x_complex = x_test.astype(np.complex128)
-
-                    for i in range(n_params):
-                        x_orig = x_complex[i]
-                        x_complex[i] = x_orig + epsilon * 1j
-
-                        # Create full parameter vector with complex perturbation
-                        x_full = np.concatenate([x_complex, x0[n_params:].astype(np.complex128)])
-                        mod.reshape(x_full)
-
-                        # Try to simulate with complex parameters
-                        try:
-                            y_hat_complex = mod.simulate(u)
-                            loss_complex = np.sum((y - y_hat_complex) ** 2)
-                            grad_complex[i] = np.imag(loss_complex) / epsilon
-                        except (TypeError, ValueError):
-                            # Model doesn't support complex parameters, fall back to finite differences
-                            raise ValueError("Model does not support complex parameters for complex step.") from None
-
-                        x_complex[i] = x_orig
-
-                    elapsed = time.perf_counter() - start
-                    complex_times.append(elapsed)
-                    complex_errors.append(float(np.linalg.norm(grad_complex - true_grad)))
-
-                results["times"]["complex"].append(np.median(complex_times))
-                results["errors"]["complex"].append(np.median(complex_errors))
-            except (TypeError, ValueError):
-                # Fall back to finite differences with small epsilon for complex step
-                for _ in range(n_runs):
-                    start = time.perf_counter()
-
-                    epsilon = 1e-20
-                    grad_complex = np.zeros_like(x_test)
-
-                    for i in range(n_params):
-                        x_perturbed = x_test.copy()
-                        x_perturbed[i] += epsilon
-
-                        x_full = np.concatenate([x_perturbed, x0[n_params:]])
-                        mod.reshape(x_full)
-                        loss_perturbed = LTIModel.SSE(y - mod.simulate(u))
-
-                        grad_complex[i] = (loss_perturbed - loss_nominal) / epsilon
-
-                    elapsed = time.perf_counter() - start
-                    complex_times.append(elapsed)
-                    complex_errors.append(float(np.linalg.norm(grad_complex - true_grad)))
-
-                results["times"]["complex"].append(np.median(complex_times))
-                results["errors"]["complex"].append(np.median(complex_errors))
-
-        return results
-
-    # Use custom loss function or default quadratic
-    if loss_function is None:
-
-        def loss_function(x: np.ndarray) -> float:
-            """Default quadratic loss for benchmarking."""
-            # Use a fixed x_true for reproducibility
-            np.random.seed(42)
-            x_true = np.random.randn(max(n_params_list))
-            return float(np.sum((x - x_true[: len(x)]) ** 2))
-
-    for n_params in n_params_list:
-        np.random.seed(42)
-        x_true = np.random.randn(n_params)
-
-        def simple_loss(x: np.ndarray, x_true=x_true) -> float:
-            """Simple quadratic loss for benchmarking."""
-            return float(np.sum((x - x_true) ** 2))
-
-        # True gradient: 2*(x - x_true)
-        x_test = np.random.randn(n_params)
-        true_grad = 2 * (x_test - x_true)
-
-        # Benchmark finite differences
-        finite_times = []
-        finite_errors = []
-
-        for _ in range(n_runs):
-            start = time.perf_counter()
-
-            # Use scipy's approx_fprime
-            grad_finite = scipy.optimize.approx_fprime(x_test, simple_loss, epsilon=1e-8)
-
-            elapsed = time.perf_counter() - start
-            finite_times.append(elapsed)
-            finite_errors.append(float(np.linalg.norm(grad_finite - true_grad)))
-
-        results["times"]["finite"].append(np.median(finite_times))
-        results["errors"]["finite"].append(np.median(finite_errors))
-
-        # Benchmark complex step
-        complex_times = []
-        complex_errors = []
-
-        for _ in range(n_runs):
-            start = time.perf_counter()
-
-            # Complex step implementation
-            epsilon = 1e-20
-            grad_complex = np.zeros_like(x_test)
-            x_complex = x_test.astype(np.complex128)
-
-            for i in range(n_params):
-                x_orig = x_complex[i]
-                x_complex[i] = x_orig + epsilon * 1j
-
-                # Evaluate loss with complex perturbation
-                loss_complex = np.sum((x_complex - x_true) ** 2)
-
-                grad_complex[i] = np.imag(loss_complex) / epsilon
-                x_complex[i] = x_orig
-
-            elapsed = time.perf_counter() - start
-            complex_times.append(elapsed)
-            complex_errors.append(float(np.linalg.norm(grad_complex - true_grad)))
-
-        results["times"]["complex"].append(np.median(complex_times))
-        results["errors"]["complex"].append(np.median(complex_errors))
-
-    return results
-
-
-def print_benchmark_results(results: dict[str, Any]) -> None:
-    """Print benchmark results in a formatted table."""
-    print("\n" + "=" * 70)
-    print("DERIVATIVE METHOD BENCHMARK RESULTS")
-    print("=" * 70)
-    print(f"{'Method':<12} {'N Params':<10} {'Time (ms)':<12} {'Error':<15}")
-    print("-" * 70)
-
-    for i, n_params in enumerate(results["n_params"]):
-        for method in results["methods"]:
-            time_ms = results["times"][method][i] * 1000
-            error = results["errors"][method][i]
-            print(f"{method:<12} {n_params:<10} {time_ms:<12.4f} {error:<15.2e}")
-
-    print("-" * 70)
-    print("\nSummary:")
-    print("- 'complex' (default): Faster AND more accurate (complex step method)")
-    print("- 'finite': Slower and less accurate (finite differences)")
-    print("\nRecommendation:")
-    print("- Use 'complex' (default) for most cases")
-    print("- Use 'finite' only for compatibility with legacy code")
-    print("=" * 70 + "\n")
